Move ReST external stuff to a sub-module
