+++ src/cmd/manager/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/halprin/email-conceal/src/external"
+	"github.com/halprin/email-conceal/src/external/rest"
 )
 
 func main() {
-	external.Rest()
+	rest.Rest()
 }
